main: blink led wakeup pattern after initializing the USB Stack
Blinking the LEDs for each channel can add up to quite a significant
delay in the initialization. For devices that cannot re-trigger USB
enumeration due to hardwired pullups on the USB D+ line this may lead
to timeouts and errors (Device not responding to setup address).

Blink the LEDs after setting up the USB Peripheral and USB stack to
workaround this issue.

Signed-off-by: Johannes Zink <opensource@johannes-zink.de>